alert-monitoring-back-web-api/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
asyncpg==0.30.0
build==1.4.4
CacheControl==0.14.4
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
cleo==2.1.0
click==8.2.1
coverage==7.13.5
crashtest==0.4.1
cryptography==46.0.7
distlib==0.4.0
dulwich==0.21.7
execnet==2.1.2
Faker==40.15.0
fastapi==0.124.4
fastjsonschema==2.21.2
filelock==3.29.0
freezegun==1.5.5
fwkpy-lib-auth==1.0.2
fwkpy-lib-core==1.0.2
fwkpy-lib-database==1.0.2
fwkpy-lib-fastapi==1.0.2
fwkpy-lib-hash==1.0.2
fwkpy-lib-utils==1.0.2
googleapis-common-protos==1.74.0
greenlet==3.2.5
grpcio==1.80.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
importlib_metadata==8.7.1
iniconfig==2.3.0
installer==0.7.0
jaraco.classes==3.4.0
jeepney==0.9.0
keyring==24.3.1
more-itertools==11.0.2
msal==1.34.0
msgpack==1.1.2
opentelemetry-api==1.37.0
opentelemetry-distro==0.58b0
opentelemetry-exporter-otlp==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-instrumentation-asgi==0.58b0
opentelemetry-instrumentation-asyncpg==0.58b0
opentelemetry-instrumentation-dbapi==0.58b0
opentelemetry-instrumentation-fastapi==0.58b0
opentelemetry-instrumentation-logging==0.58b0
opentelemetry-instrumentation-psycopg==0.58b0
opentelemetry-instrumentation-sqlalchemy==0.58b0
opentelemetry-instrumentation-system-metrics==0.58b0
opentelemetry-propagator-b3==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opentelemetry-util-http==0.58b0
packaging==26.2
pexpect==4.9.0
pkginfo==1.12.1.2
platformdirs==4.9.6
pluggy==1.6.0
poetry==1.8.5
poetry-core==1.9.1
poetry-plugin-export==1.8.0
protobuf==6.33.6
psutil==7.0.0
psycopg==3.2.13
psycopg-binary==3.2.13
ptyprocess==0.7.0
pyaml_env==1.2.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyfakefs==5.7.4
PyJWT==2.12.1
pyproject_hooks==1.2.0
pytest==8.3.5
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-env==1.1.5
pytest-faker==2.0.0
pytest-mock==3.14.0
pytest-xdist==3.6.1
pytest_freezer==0.4.9
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-i18n==0.3.9
python-multipart==0.0.27
PyYAML==6.0.3
RapidFuzz==3.14.5
requests==2.33.1
requests-toolbelt==1.0.0
SecretStorage==3.5.0
shellingham==1.5.4
six==1.17.0
SQLAlchemy==2.0.49
starlette==0.49.3
tomli==2.2.1
tomlkit==0.14.0
trove-classifiers==2026.1.14.14
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.34.3
virtualenv==20.39.1
wrapt==1.17.3
zipp==3.23.1
